+++ src/App.jsx
@@ -2,12 +2,14 @@ import "./App.css";
 import ContactUs from "./components/ContactUs";
 import Footer from "./components/Footer";
 import Navbar from "./components/Navbar";
+import Hero from "./components/Hero";
 
 function App() {
   return (
     <>
       <div className="App">
         <Navbar />
+        <Hero/>
         <ContactUs />
         <Footer />
       </div>
