+++ src/styles/Navbar.module.css
@@ -7,7 +7,7 @@
     padding: 20px 80px ;
     box-shadow:0px 0px 29px -6px rgba(0, 0, 0, 1) ;
     position: fixed; 
-    z-index:  1;
+    z-index:  1000;
     right: 0;
     left: 0; 
     top:0;
